+++ docs/build.sh
@@ -1,10 +1,12 @@
 make clean
 
 rm -rf source/auto_tutorials/
+
 if [ "$1" = "wipe" ]; then
     echo "Wiping all .fits files under ./source/tutorials/..."
     find source/tutorials/ -type f -name "*.fits" -delete
 fi
+
 if [ "$1" = "wipe" ]; then
     echo "Wiping all .fits.gz files under ./source/tutorials/..."
     find source/tutorials/ -type f -name "*.fits.gz" -delete